1. 创建新的module spring_basic. 并将现有代码移入该module
