+++ spring_basic/src/main/java/com/tinyv/demo/global/config/properties/LocalCacheProperties.java
