+++ spring_basic/src/main/java/com/tinyv/demo/global/constant/ConstGlobal.java
