--- src/main/resources/application-dev.properties
@@ -1,104 +1,0 @@
-spring.application.name=Spring2X
-
-
-##log config
-
-#在application.properties文件中引入日志配置文件
-#=====================================  log  =============================
-logging.config=classpath:logback-boot.xml
-
-
-##spring cache config
-spring.cache.caffeine.initialCapacity=50
-spring.cache.caffeine.maximumSize=500
-#基于权重来进行驱逐，不可以和maximumSize一起使用
-spring.cache.caffeine.maximumWeight=200
-spring.cache.caffeine.expireAfterAccess=2
-spring.cache.caffeine.expireAfterWrite=10
-spring.cache.caffeine.refreshAfterWrite=1
-spring.cache.caffeine.weakKeys=false
-spring.cache.caffeine.weakValues=false
-spring.cache.caffeine.softValues=false
-spring.cache.caffeine.recordStats=true
-
-##Mysql config
-
-
-####mysql
-spring.datasource.username=root
-spring.datasource.password=Passw0rd_R
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url=jdbc:mysql://122.51.234.242:3306/spring2X
-#连接池启动时创建的初始化连接数量（默认值为0）
-spring.datasource.initialSize=5
-#连接池的最大数据库连接数。设为0表示无限制
-spring.datasource.maxActive=20
-#最大空闲连接数--连接池中容许保持空闲状态的最大连接数量,超过的空闲连接将被释放,如果设置为负数表示不限制
-spring.datasource.maxIdle=10
-#最小空闲连接数--连接池中容许保持空闲状态的最小连接数量,低于这个数量将创建新的连接,如果设置为0则不创建
-spring.datasource.minIdle=5
-#最大建立连接等待时间。如果超过此时间将接到异常。设为-1表示无限制。
-spring.datasource.maxWait=60000
-#超时时间(以秒数为单位)
-spring.datasource.removeAbandonedTimeout=180
-#是否自动回收超时连接
-spring.datasource.removeAbandoned=true
-spring.datasource.type =com.alibaba.druid.pool.DruidDataSource
-
-#Mybatis
-mybatis.type-aliases-package=com.tinyv.demo.business.bean
-mybatis.mapper-locations=classpath:mapper/*.xml
-# 开启驼峰命名转换，如：Table(create_time) -> Entity(createTime)。不需要我们关心怎么进行字段匹配，mybatis会自动识别`大写字母与下划线`
-mybatis.configuration.map-underscore-to-camel-case=true
-mybatis.logging.level=INFO
-
-### RocketMQ
-rocketmq.namesrvAddr = 127.0.0.1:9876
-rocketmq.groupName=${spring.application.name}
-rocketmq.topics=topicDemo
-rocketmq.tags=tag1;tag2
-
-###producer
-#该应用是否启用生产者
-rocketmq.producer.isOnOff=on
-#发送同一类消息的设置为同一个group，保证唯一,默认不需要设置，rocketmq会使用ip@pid(pid代表jvm名字)作为唯一标示
-rocketmq.producer.groupName=${spring.application.name}
-#mq的nameserver地址
-rocketmq.producer.namesrvAddr=${rocketmq.namesrvAddr}
-#消息最大长度 默认1024*4(4M)
-rocketmq.producer.maxMessageSize=4096
-#发送消息超时时间,默认3000
-rocketmq.producer.sendMsgTimeout=3000
-#发送消息失败重试次数，默认2
-rocketmq.producer.retryTimesWhenSendFailed=2
-
-###consumer
-##该应用是否启用消费者
-rocketmq.consumer.isOnOff=on
-rocketmq.consumer.groupName=${spring.application.name}
-#mq的nameserver地址
-rocketmq.consumer.namesrvAddr=${rocketmq.namesrvAddr}
-#该消费者订阅的主题和tags("*"号表示订阅该主题下所有的tags),格式：topic~tag1||tag2||tag3;topic2~*;
-rocketmq.consumer.topics=topicDemo~*;
-rocketmq.consumer.consumeThreadMin=20
-rocketmq.consumer.consumeThreadMax=64
-#设置一次消费消息的条数，默认为1条
-rocketmq.consumer.consumeMessageBatchMaxSize=1
-
-
-### redis
-# 连接池属性
-spring.redis.minIdle=5
-spring.redis.maxIdle=20
-spring.redis.maxActive=100
-spring.redis.maxWait=20000
-# 服务器信息
-spring.redis.host=122.51.234.242
-spring.redis.port=6379
-spring.redis.password=password
-# 连接超时时间
-spring.redis.timeout=10000
-
-
-
-
